pre deployment
