+++ package.json
@@ -31,7 +31,8 @@
     "passport-facebook": "^3.0.0",
     "passport-google-oauth20": "^2.0.0",
     "pg": "^8.11.5",
-    "redis": "^5.8.2"
+    "redis": "^5.8.2",
+    "safe-stable-stringify": "^2.5.0"
   },
   "devDependencies": {
     "nodemon": "^3.0.1"
