+++ server.js
@@ -1,1161 +1,51 @@
 // server.js
-const { createClient } = require('redis');
 const express = require('express');
 const cors = require('cors');
-const bcrypt = require('bcrypt');
-const jwt = require('jsonwebtoken');
-const passport = require('passport');
-const FacebookStrategy = require('passport-facebook').Strategy;
-const GoogleStrategy = require('passport-google-oauth20').Strategy;
 const session = require('express-session');
+const passport = require('passport');
 const RedisStore = require('connect-redis').default;
-const { knex: db, redis } = require('./db');
 
+const { redis } = require('./db');
+const { PORT } = require('./config/env');
+require('./config/passport');
 
-const redisClient = createClient({
-    socket: {
-        host: 'redis', // Use the Docker service name
-        port: 6379
-    }
-});
+const { authenticateToken } = require('./middleware/auth');
+const { notFoundHandler, errorHandler } = require('./middleware/errorHandlers');
 
-
-redisClient.on('error', (err) => console.error('Redis Client Error', err));
-
-async function initRedis() {
-    try {
-        await redisClient.connect();
-        console.log('Redis connected');
-
-        // Now it's safe to use Redis
-        await redisClient.ping();
-        console.log('Ping successful');
-    } catch (err) {
-        console.error('Redis connection failed:', err);
-    }
-}
-
-initRedis();
+const healthRoutes = require('./routes/healthRoutes');
+const flagsRoutes = require('./routes/flagsRoutes');
+const authRoutes = require('./routes/authRoutes');
+const userRoutes = require('./routes/userRoutes');
+const gameRoutes = require('./routes/gameRoutes');
+const leaderboardRoutes = require('./routes/leaderboardRoutes');
 
 const app = express();
-const PORT = process.env.PORT || 3000;
-const JWT_SECRET = process.env.JWT_SECRET || 'minemaster_secret_key_2024';
-if (!process.env.JWT_SECRET) {
-    console.warn('Warning: Using default JWT_SECRET. Set process.env.JWT_SECRET in production.');
-}
 
-// OAuth Configuration
-const FACEBOOK_APP_ID = process.env.FACEBOOK_APP_ID || 'your_facebook_app_id';
-const FACEBOOK_APP_SECRET = process.env.FACEBOOK_APP_SECRET || 'your_facebook_app_secret';
-const GOOGLE_CLIENT_ID = process.env.GOOGLE_CLIENT_ID || 'your_google_client_id';
-const GOOGLE_CLIENT_SECRET = process.env.GOOGLE_CLIENT_SECRET || 'your_google_client_secret';
-const BASE_URL = process.env.BASE_URL || `http://localhost:${PORT}`;
-
-// Middleware
-app.use(cors({
-    origin: true,
-    credentials: true
-}));
+app.use(cors({ origin: true, credentials: true }));
 app.use(express.json({ limit: '10mb' }));
 
 app.use(session({
-    store: new RedisStore({
-        client: redis,
-        prefix: "myapp:"
-    }),
+    store: new RedisStore({ client: redis, prefix: "myapp:" }),
     secret: process.env.SESSION_SECRET || 'your_session_secret',
     resave: false,
     saveUninitialized: false,
-    cookie: {
-        secure: process.env.NODE_ENV === 'production',
-        maxAge: 24 * 60 * 60 * 1000 // 24 hours
-    }
+    cookie: { secure: process.env.NODE_ENV === 'production', maxAge: 24 * 60 * 60 * 1000 },
 }));
 
 app.use(passport.initialize());
 app.use(passport.session());
 
-// Simple startup checks (non-blocking)
-const MAX_RETRIES = 10;
-let retries = 0;
-
-async function connectToPostgres() {
-    while (retries < MAX_RETRIES) {
-        try {
-            await db.raw('SELECT 1'); // or your preferred connection test
-            console.log('Postgres connection OK');
-            break;
-        } catch (err) {
-            if (err.code === '57P03') {
-                console.log('Postgres is starting up, retrying...');
-                retries++;
-                await new Promise(res => setTimeout(res, 1000)); // wait 1 second
-            } else {
-                console.error('Postgres connection error', err);
-                break;
-            }
-        }
-    }
-}
-// db.raw('SELECT 1').then(() => console.log('Postgres connection OK')).catch(err => console.error('Postgres connection error', err));
-redis.ping().then(() => console.log('Redis connection OK')).catch(err => console.error('Redis connection error', err));
-
-// Passport serialize/deserialize using Knex
-passport.serializeUser((user, done) => {
-    done(null, user.id);
-});
-
-passport.deserializeUser((id, done) => {
-    db('users').where({ id }).first()
-        .then(user => done(null, user))
-        .catch(err => done(err));
-});
-
-// // Passport strategies
-// passport.use(new FacebookStrategy({
-//     clientID: FACEBOOK_APP_ID,
-//     clientSecret: FACEBOOK_APP_SECRET,
-//     callbackURL: `${BASE_URL}/api/auth/facebook/callback`,
-//     profileFields: ['id', 'emails', 'name', 'picture']
-// }, (accessToken, refreshToken, profile, done) => {
-//     try {
-//         const facebookId = profile.id;
-//         const email = profile.emails && profile.emails[0] ? profile.emails[0].value : null;
-//         const firstName = (profile.name && profile.name.givenName) || '';
-//         const lastName = (profile.name && profile.name.familyName) || '';
-//         const usernameBase = `${firstName}${lastName}`.toLowerCase().replace(/[^a-z0-9]/g, '') || `facebook_user_${facebookId}`;
-//         const profilePicture = profile.photos && profile.photos[0] ? profile.photos[0].value : null;
-
-//         // Find user by oauth provider/id
-//         db('users').where({ oauth_provider: 'facebook', oauth_id: facebookId }).first()
-//             .then(existingUser => {
-//                 if (existingUser) return done(null, existingUser);
-
-//                 // If email exists, link account
-//                 if (email) {
-//                     return db('users').where({ email }).first()
-//                         .then(emailUser => {
-//                             if (emailUser) {
-//                                 return db('users').where({ id: emailUser.id }).update({
-//                                     oauth_provider: 'facebook',
-//                                     oauth_id: facebookId,
-//                                     // profile_picture: profilePicture,
-//                                     updated_at: db.fn.now()
-//                                 }).then(() => {
-//                                     return db('users').where({ id: emailUser.id }).first().then(u => done(null, u));
-//                                 });
-//                             }
-
-//                             // create new user with email
-//                             return createOauthUser(usernameBase, email, 'facebook', facebookId, profilePicture, done);
-//                         });
-//                 }
-
-//                 // create new user without verified email
-//                 return createOauthUser(usernameBase, `${usernameBase}@facebook.local`, 'facebook', facebookId, profilePicture, done);
-//             })
-//             .catch(err => done(err));
-//     } catch (err) {
-//         done(err);
-//     }
-// }));
-
-passport.use(new GoogleStrategy({
-    clientID: GOOGLE_CLIENT_ID,
-    clientSecret: GOOGLE_CLIENT_SECRET,
-    callbackURL: `${BASE_URL}/api/auth/google/callback`
-}, (accessToken, refreshToken, profile, done) => {
-    try {
-        const googleId = profile.id;
-        const email = profile.emails && profile.emails[0] ? profile.emails[0].value : null;
-        const firstName = (profile.name && profile.name.givenName) || '';
-        const lastName = (profile.name && profile.name.familyName) || '';
-        const usernameBase = `${firstName}${lastName}`.toLowerCase().replace(/[^a-z0-9]/g, '') || `google_user_${googleId}`;
-        const profilePicture = profile.photos && profile.photos[0] ? profile.photos[0].value : null;
-
-        db('users').where({ oauth_provider: 'google', oauth_id: googleId }).first()
-            .then(existingUser => {
-                if (existingUser) return done(null, existingUser);
-
-                if (email) {
-                    return db('users').where({ email }).first()
-                        .then(emailUser => {
-                            if (emailUser) {
-                                return db('users').where({ id: emailUser.id }).update({
-                                    oauth_provider: 'google',
-                                    oauth_id: googleId,
-                                    // profile_picture: profilePicture,
-                                    updated_at: db.fn.now()
-                                }).then(() => db('users').where({ id: emailUser.id }).first().then(u => done(null, u)));
-                            }
-
-                            return createOauthUser(usernameBase, email, 'google', googleId, profilePicture, done);
-                        });
-                }
-
-                return createOauthUser(usernameBase, `${usernameBase}@google.local`, 'google', googleId, profilePicture, done);
-            })
-            .catch(err => done(err));
-    } catch (err) {
-        done(err);
-    }
-}));
-
-// Helper to create OAuth users (ensures unique username)
-function createOauthUser(usernameBase, email, provider, oauthId, profilePicture, done) {
-    function tryUsername(candidate, attempt = 0) {
-        const testName = attempt === 0 ? candidate : `${candidate}${attempt}`;
-        return db('users').where({ username: testName }).first()
-            .then(existing => {
-                if (existing) return tryUsername(candidate, attempt + 1);
-                return db('users').insert({
-                    username: testName,
-                    email,
-                    oauth_provider: provider,
-                    oauth_id: oauthId,
-                    // profile_picture: profilePicture,
-                    country_flag: 'international',
-                    created_at: db.fn.now(),
-                    updated_at: db.fn.now()
-                }).returning('id').then(rows => ({ game_id: rows[0].id || rows[0] }))
-            });
-    }
-
-    return tryUsername(usernameBase);
-}
-
-// JWT Authentication middleware
-const authenticateToken = (req, res, next) => {
-    const authHeader = req.headers['authorization'];
-    const token = authHeader && authHeader.split(' ')[1];
-
-    if (!token) return res.status(401).json({ error: 'Access token required' });
-
-    jwt.verify(token, JWT_SECRET, (err, user) => {
-        if (err) return res.status(403).json({ error: 'Invalid token' });
-        req.user = user;
-        next();
-    });
-};
-
-// Available country flags
-const AVAILABLE_FLAGS = [
-    'international', 'af', 'al', 'dz', 'ad', 'ao', 'ag', 'ar', 'am', 'au', 'at', 'az',
-    'bs', 'bh', 'bd', 'bb', 'by', 'be', 'bz', 'bj', 'bt', 'bo', 'ba', 'bw', 'br', 'bn',
-    'bg', 'bf', 'bi', 'kh', 'cm', 'ca', 'cv', 'cf', 'td', 'cl', 'cn', 'co', 'km', 'cg',
-    'cr', 'hr', 'cu', 'cy', 'cz', 'dk', 'dj', 'dm', 'do', 'tl', 'ec', 'eg', 'sv', 'gq',
-    'er', 'ee', 'sz', 'et', 'fj', 'fi', 'fr', 'ga', 'gm', 'ge', 'de', 'gh', 'gr', 'gd',
-    'gt', 'gn', 'gw', 'gy', 'ht', 'hn', 'hu', 'is', 'in', 'id', 'ir', 'iq', 'ie', 'il',
-    'it', 'ci', 'jm', 'jp', 'jo', 'kz', 'ke', 'ki', 'kw', 'kg', 'la', 'lv', 'lb', 'ls',
-    'lr', 'ly', 'li', 'lt', 'lu', 'mg', 'mw', 'my', 'mv', 'ml', 'mt', 'mh', 'mr', 'mu',
-    'mx', 'fm', 'md', 'mc', 'mn', 'me', 'ma', 'mz', 'mm', 'na', 'nr', 'np', 'nl', 'nz',
-    'ni', 'ne', 'ng', 'kp', 'mk', 'no', 'om', 'pk', 'pw', 'ps', 'pa', 'pg', 'py', 'pe',
-    'ph', 'pl', 'pt', 'qa', 'ro', 'ru', 'rw', 'kn', 'lc', 'vc', 'ws', 'sm', 'st', 'sa',
-    'sn', 'rs', 'sc', 'sl', 'sg', 'sk', 'si', 'sb', 'so', 'za', 'kr', 'ss', 'es', 'lk',
-    'sd', 'sr', 'se', 'ch', 'sy', 'tw', 'tj', 'tz', 'th', 'tg', 'to', 'tt', 'tn', 'tr',
-    'tm', 'tv', 'ug', 'ua', 'ae', 'gb', 'us', 'uy', 'uz', 'vu', 'va', 've', 'vn', 'ye',
-    'zm', 'zw'
-];
-// --- Routes ---
-
-// Health
-app.get('/health', (req, res) => res.json({ status: 'OK', message: 'Mine Master API is running' }));
-
-// Debug: only enabled when DEBUG=true
-if (process.env.DEBUG === 'true') {
-    app.get('/api/debug/tables', (req, res) => {
-        const results = {};
-        db.select('*').from('users').limit(500).then(users => {
-            results.users = users;
-            return db.select('*').from('scores').limit(500);
-        }).then(scores => {
-            results.scores = scores;
-            return db.select('*').from('game_states').limit(500);
-        }).then(game_states => {
-            results.game_states = game_states;
-            res.json(results);
-        }).catch(err => res.status(500).json({ error: err.message }));
-    });
-}
-
-// Flags
-app.get('/api/flags', (req, res) => res.json({ flags: AVAILABLE_FLAGS }));
-
-// Register
-app.post('/api/auth/register', (req, res) => {
-    const { username, email, password, country_flag = 'international' } = req.body;
-
-    if (!username || !email || !password) return res.status(400).json({ error: 'Username, email, and password are required' });
-    if (password.length < 6) return res.status(400).json({ error: 'Password must be at least 6 characters long' });
-    if (!AVAILABLE_FLAGS.includes(country_flag)) return res.status(400).json({ error: 'Invalid country flag' });
-
-    bcrypt.hash(password, 10).then(passwordHash => {
-        return db('users').insert({
-            username,
-            email,
-            password_hash: passwordHash,
-            country_flag,
-            created_at: db.fn.now(),
-            updated_at: db.fn.now()
-        }).returning(['id', 'username', 'email']);
-    }).then(rows => {
-        const user = rows[0];
-        const token = jwt.sign({ id: user.id, username: user.username }, JWT_SECRET, { expiresIn: '24h' });
-        res.status(201).json({
-            message: 'User created successfully',
-            user: { id: user.id, username: user.username, email: user.email, country_flag, auth_method: 'traditional' },
-            token
-        });
-    }).catch(err => {
-        if (err && err.constraint && (err.constraint.includes('users_username_unique') || err.constraint.includes('users_email_unique'))) {
-            return res.status(409).json({ error: 'Username or email already exists' });
-        }
-        console.error(err);
-        res.status(500).json({ error: 'Failed to create user' });
-    });
-});
-
-// Login
-app.post('/api/auth/login', (req, res) => {
-    const { username, password } = req.body;
-    if (!username || !password) return res.status(400).json({ error: 'Username and password are required' });
-
-    db('users').where({ username }).andWhere('oauth_provider', null).first()
-        .then(user => {
-            if (!user) {
-                return res.status(404).json({ error: 'Account not found' });
-            }
-
-            if (!user.password_hash) {
-                return res.status(401).json({ error: 'Invalid credentials' });
-            }
-
-            return bcrypt.compare(password, user.password_hash).then(ok => {
-                if (!ok) return res.status(401).json({ error: 'Invalid credentials' });
-                const token = jwt.sign({ id: user.id, username: user.username }, JWT_SECRET, { expiresIn: '24h' });
-                res.json({
-                    message: 'Login successful',
-                    user: { id: user.id, username: user.username, email: user.email, country_flag: user.country_flag, auth_method: 'traditional' },
-                    token
-                });
-            });
-        }).catch(err => {
-            console.error(err);
-            res.status(500).json({ error: 'Server error' });
-        });
-});
-// Reset Password with detailed logging
-app.post('/api/auth/reset-password', (req, res) => {
-    const { username, email, new_password } = req.body;
-
-    console.log('=== PASSWORD RESET REQUEST ===');
-    console.log('Raw body:', req.body);
-    console.log('Username received:', username);
-    console.log('Email received:', email);
-    console.log('Password length:', new_password ? new_password.length : 0);
-
-    // Validation
-    if (!username || !email || !new_password) {
-        console.log('❌ Validation failed: Missing fields');
-        return res.status(400).json({ error: 'Username, email, and new password are required' });
-    }
-
-    if (new_password.length < 6) {
-        console.log('❌ Validation failed: Password too short');
-        return res.status(400).json({ error: 'Password must be at least 6 characters long' });
-    }
-
-    const trimmedUsername = username.trim();
-    const trimmedEmail = email.trim().toLowerCase();
-
-    console.log('After trimming:');
-    console.log('  Username:', `"${trimmedUsername}"`);
-    console.log('  Email:', `"${trimmedEmail}"`);
-
-    // Find user by username AND email (both must match for security)
-    db('users')
-        .where({ username: trimmedUsername, email: trimmedEmail })
-        .first()
-        .then(user => {
-            console.log('Database query result:', user ? 'User found' : 'No user found');
-
-            if (user) {
-                console.log('Found user:', {
-                    id: user.id,
-                    username: user.username,
-                    email: user.email
-                });
-            } else {
-                console.log('❌ No matching user in database');
-                console.log('Searching for all users with this username:');
-
-                // Debug: Check if username exists at all
-                return db('users')
-                    .where({ username: trimmedUsername })
-                    .first()
-                    .then(userByUsername => {
-                        if (userByUsername) {
-                            console.log('Found user by username only:', {
-                                username: userByUsername.username,
-                                email: userByUsername.email
-                            });
-                            console.log('⚠️ Username exists but email does not match!');
-                            console.log(`  Provided email: "${trimmedEmail}"`);
-                            console.log(`  Actual email: "${userByUsername.email}"`);
-                        } else {
-                            console.log('❌ Username does not exist in database at all');
-                        }
-
-                        return res.status(404).json({
-                            error: 'No account found with this username and email'
-                        });
-                    });
-            }
-
-            // Hash the new password
-            return bcrypt.hash(new_password, 10).then(passwordHash => {
-                console.log('✅ Password hashed successfully');
-
-                // Update the password in the database
-                return db('users')
-                    .where({ id: user.id })
-                    .update({
-                        password_hash: passwordHash,
-                        updated_at: db.fn.now()
-                    })
-                    .then(() => {
-                        console.log('✅ Password updated successfully for user:', user.username);
-                        res.json({
-                            message: 'Password reset successfully',
-                            success: true
-                        });
-                    });
-            });
-        })
-        .catch(err => {
-            console.error('❌ Password reset error:', err);
-            res.status(500).json({ error: 'Failed to reset password' });
-        });
-});
-
-app.post('/api/auth/facebook', async (req, res) => {
-    console.log('\n=== FACEBOOK AUTH REQUEST START ===');
-    console.log('📩 Request received at:', new Date().toISOString());
-    console.log('📩 Headers:', JSON.stringify(req.headers, null, 2));
-    console.log('📩 Body:', JSON.stringify(req.body, null, 2));
-
-    const { facebook_id, name, email } = req.body;
-
-    console.log('📝 Extracted values:');
-    console.log('  - Facebook ID:', facebook_id, '(type:', typeof facebook_id, ')');
-    console.log('  - Name:', name, '(type:', typeof name, ')');
-    console.log('  - Email:', email, '(type:', typeof email, ')');
-
-    // Validation
-    if (!facebook_id || !name) {
-        console.log('❌ Validation failed: Missing required fields');
-        return res.status(400).json({ error: 'Facebook ID and name are required' });
-    }
-
-    try {
-        // --------------------------------------------------
-        // 1) EXISTING USER BY OAUTH
-        // --------------------------------------------------
-        console.log('🔍 Searching for existing Facebook user by oauth...');
-        let user = await db('users')
-            .where({ oauth_provider: 'facebook', oauth_id: facebook_id })
-            .first();
-
-        console.log('📊 OAuth query result:', user ? 'User FOUND' : 'User NOT FOUND');
-
-        // --------------------------------------------------
-        // 2) IF NOT FOUND, TRY LINK BY EMAIL (IF PROVIDED)
-        //    (handles case where user signed up with email/password before)
-        // --------------------------------------------------
-        if (!user && email) {
-            console.log('🔗 Trying to link existing account by email:', email);
-
-            const emailUser = await db('users').where({ email }).first();
-
-            if (emailUser) {
-                console.log('✅ Found existing user by email. Linking Facebook oauth to this user:', {
-                    id: emailUser.id,
-                    username: emailUser.username,
-                    email: emailUser.email,
-                });
-
-                await db('users')
-                    .where({ id: emailUser.id })
-                    .update({
-                        oauth_provider: 'facebook',
-                        oauth_id: facebook_id,
-                        updated_at: db.fn.now(),
-                    });
-
-                user = await db('users')
-                    .select('id', 'username', 'email', 'country_flag')
-                    .where({ id: emailUser.id })
-                    .first();
-            } else {
-                console.log('ℹ️ No existing user found by email to link.');
-            }
-        }
-
-        // --------------------------------------------------
-        // 3) IF STILL NOT FOUND, CREATE NEW USER
-        // --------------------------------------------------
-        if (!user) {
-            console.log('🆕 New Facebook user — creating account...');
-
-            const baseUsername = String(name).toLowerCase().replace(/[^a-z0-9]/g, '') || 'user';
-
-            // Try a few times to avoid username collisions
-            let createdUser = null;
-
-            for (let attempt = 1; attempt <= 5; attempt++) {
-                const username = `${baseUsername}${Math.floor(Math.random() * 100000)}`;
-
-                const newUser = {
-                    username,
-                    email: email || `${facebook_id}@facebook.temp`,
-                    oauth_provider: 'facebook',
-                    oauth_id: facebook_id,
-                    country_flag: 'international',
-                    created_at: db.fn.now(),
-                    updated_at: db.fn.now(),
-                };
-
-                // IMPORTANT: do NOT JSON.stringify(newUser) because db.fn.now() is not JSON-serializable
-                console.log('💾 Inserting new user:', {
-                    ...newUser,
-                    created_at: 'now()',
-                    updated_at: 'now()',
-                });
-
-                try {
-                    const inserted = await db('users').insert(newUser);
-
-                    // Works across DBs: inserted id may be number or [number]
-                    const insertedId = Array.isArray(inserted) ? inserted[0] : inserted;
-
-                    createdUser = await db('users')
-                        .select('id', 'username', 'email', 'country_flag')
-                        .where({ id: insertedId })
-                        .first();
-
-                    if (createdUser) break;
-                } catch (e) {
-                    // If username/email collision, retry; otherwise throw
-                    const msg = (e && e.message) ? e.message : String(e);
-
-                    const isUniqueViolation =
-                        msg.toLowerCase().includes('unique') ||
-                        msg.toLowerCase().includes('duplicate') ||
-                        msg.toLowerCase().includes('constraint');
-
-                    console.error(`❌ Insert attempt ${attempt} failed:`, msg);
-
-                    // If email is already taken (and we couldn't link), surface a clean error
-                    if (email && msg.toLowerCase().includes('email')) {
-                        return res.status(409).json({
-                            error:
-                                'An account with this email already exists. Please login with your existing method, then link Facebook, or contact support.',
-                        });
-                    }
-
-                    if (!isUniqueViolation || attempt === 5) throw e;
-                }
-            }
-
-            if (!createdUser) {
-                throw new Error('Failed to create user after multiple attempts');
-            }
-
-            user = createdUser;
-
-            console.log('✅ New user created:', user);
-
-            const token = jwt.sign({ id: user.id, username: user.username }, JWT_SECRET, {
-                expiresIn: '24h',
-            });
-
-            const response = {
-                message: 'User created successfully via Facebook',
-                user: {
-                    id: user.id,
-                    username: user.username,
-                    email: user.email,
-                    country_flag: user.country_flag,
-                    auth_provider: 'facebook',
-                },
-                token,
-            };
-
-            console.log('📤 Sending response (201):', JSON.stringify(response, null, 2));
-            console.log('=== FACEBOOK AUTH REQUEST END (NEW USER) ===\n');
-
-            return res.status(201).json(response);
-        }
-
-        // --------------------------------------------------
-        // 4) EXISTING (OR LINKED) USER LOGIN
-        // --------------------------------------------------
-        console.log('👤 Existing/Linked user details:', {
-            id: user.id,
-            username: user.username,
-            email: user.email,
-        });
-
-        await db('users').where({ id: user.id }).update({ updated_at: db.fn.now() });
-
-        console.log('🔑 Generating JWT token...');
-
-        const token = jwt.sign({ id: user.id, username: user.username }, JWT_SECRET, {
-            expiresIn: '24h',
-        });
-
-        const response = {
-            message: 'Login successful',
-            user: {
-                id: user.id,
-                username: user.username,
-                email: user.email,
-                country_flag: user.country_flag,
-                auth_provider: 'facebook',
-            },
-            token,
-        };
-
-        console.log('📤 Sending response:', JSON.stringify(response, null, 2));
-        console.log('=== FACEBOOK AUTH REQUEST END (SUCCESS) ===\n');
-
-        return res.json(response);
-    } catch (err) {
-        // Safer logging (avoid circular JSON problems)
-        console.error('❌ FACEBOOK AUTH ERROR:', {
-            message: err?.message,
-            stack: err?.stack,
-        });
-        console.log('=== FACEBOOK AUTH REQUEST END (ERROR) ===\n');
-
-        return res.status(500).json({
-            error: 'Failed to authenticate with Facebook',
-            details: err?.message || 'Unknown error',
-        });
-    }
-});
-
-
-
-app.get('/api/auth/google', passport.authenticate('google', { scope: ['profile', 'email'] }));
-app.get('/api/auth/google/callback',
-    passport.authenticate('google', { failureRedirect: '/login' }),
-    (req, res) => {
-        const token = jwt.sign({ id: req.user.id, username: req.user.username }, JWT_SECRET, { expiresIn: '24h' });
-        res.redirect(`myapp://auth/success?token=${token}&user=${encodeURIComponent(JSON.stringify({
-            id: req.user.id, username: req.user.username, email: req.user.email, country_flag: req.user.country_flag, auth_method: 'google'
-        }))}`);
-    }
-);
-
-// OAuth Status check
-app.get('/api/auth/oauth/status/:provider', (req, res) => {
-    const { provider } = req.params;
-    const { oauth_id } = req.query;
-    if (!oauth_id) return res.status(400).json({ error: 'OAuth ID required' });
-
-    db('users').where({ oauth_provider: provider, oauth_id }).first()
-        .then(user => {
-            if (!user) return res.status(404).json({ error: 'User not found' });
-            const token = jwt.sign({ id: user.id, username: user.username }, JWT_SECRET, { expiresIn: '24h' });
-            res.json({
-                message: 'OAuth login successful',
-                user: { id: user.id, username: user.username, email: user.email, country_flag: user.country_flag, auth_method: provider },
-                token
-            });
-        }).catch(err => res.status(500).json({ error: 'Server error' }));
-});
-
-// Profile
-app.get('/api/user/profile', authenticateToken, (req, res) => {
-    db('users').where({ id: req.user.id }).select('id', 'username', 'email', 'country_flag', 'oauth_provider', 'created_at').first()
-        .then(user => {
-            if (!user) return res.status(404).json({ error: 'User not found' });
-            res.json({ ...user, auth_method: user.oauth_provider || 'traditional' });
-        }).catch(err => res.status(500).json({ error: 'Server error' }));
-});
-
-app.put('/api/user/profile', authenticateToken, async (req, res) => {
-    const { country_flag, username, current_password, new_password } = req.body;
-    console.log('📥 Profile update request:', {
-        user_id: req.user.id,
-        body: req.body
-    });
-
-    try {
-        const updates = {};
-
-        // Country flag update
-        if (country_flag !== undefined) {
-            if (!AVAILABLE_FLAGS.includes(country_flag)) {
-                return res.status(400).json({ error: 'Invalid country flag' });
-            }
-            updates.country_flag = country_flag;
-        }
-
-        // Username update
-        if (username !== undefined) {
-            if (!username || username.trim().length === 0) {
-                return res.status(400).json({ error: 'Username cannot be empty' });
-            }
-
-            // Check if username is already taken by another user
-            const existing = await db('users')
-                .where({ username: username.trim() })
-                .whereNot({ id: req.user.id })
-                .first();
-
-            if (existing) {
-                return res.status(409).json({ error: 'Username already taken' });
-            }
-            updates.username = username.trim();
-        }
-
-        // Password update (requires current password verification)
-        if (new_password !== undefined) {
-            if (!current_password) {
-                return res.status(400).json({ error: 'Current password required to change password' });
-            }
-
-            const user = await db('users').where({ id: req.user.id }).first();
-
-            // Check if user has OAuth login (no password set)
-            if (user.oauth_provider) {
-                return res.status(400).json({
-                    error: 'Cannot set password for OAuth accounts'
-                });
-            }
-
-            if (!user.password_hash) {
-                return res.status(400).json({
-                    error: 'No password set for this account'
-                });
-            }
-
-            // Verify current password
-            const validPassword = await bcrypt.compare(current_password, user.password_hash);
-            if (!validPassword) {
-                return res.status(401).json({ error: 'Current password is incorrect' });
-            }
-
-            if (new_password.length < 6) {
-                return res.status(400).json({
-                    error: 'New password must be at least 6 characters'
-                });
-            }
-
-            updates.password_hash = await bcrypt.hash(new_password, 10);
-        }
-
-        if (Object.keys(updates).length === 0) {
-            return res.status(400).json({ error: 'No valid fields to update' });
-        }
-
-        updates.updated_at = db.fn.now();
-
-        await db('users').where({ id: req.user.id }).update(updates);
-
-        // Get updated user info to return
-        const updatedUser = await db('users')
-            .where({ id: req.user.id })
-            .select('id', 'username', 'email', 'country_flag')
-            .first();
-
-        res.json({
-            message: 'Profile updated successfully',
-            user: updatedUser,
-            updated_fields: Object.keys(updates).filter(key => key !== 'updated_at' && key !== 'password_hash')
-        });
-    } catch (err) {
-        console.error('Profile update error:', err);
-        res.status(500).json({ error: 'Failed to update profile' });
-    }
-});
-
-// Start new game
-app.post('/api/game/start', authenticateToken, (req, res) => {
-    const { mine_count, mine_positions, level_id, hints, streak } = req.body;
-
-    if (mine_count === undefined || !Array.isArray(mine_positions)) {
-        return res.status(400).json({ error: 'Grid dimensions, mine count, and mine positions are required' });
-    }
-    if (mine_positions.length !== mine_count) {
-        return res.status(400).json({ error: 'Mine positions count must match mine_count' });
-    }
-
-    // Remove any existing active game for this user
-    db('game_states').where({ user_id: req.user.id, game_status: 'playing' }).del()
-        .then(() => {
-            return db('game_states')
-                .insert({
-                    user_id: req.user.id,
-                    mine_count,
-                    level_id,
-                    mine_positions: JSON.stringify(mine_positions),
-                    revealed_cells: JSON.stringify([]),
-                    flagged_cells: JSON.stringify([]),
-                    hints: hints || 0,
-                    streak: streak || 0,
-                    game_status: 'playing',
-                    created_at: db.fn.now(),
-                    updated_at: db.fn.now()
-                })
-                .returning('id');
-        })
-        .then(rows => {
-            // rows[0] could be { id: 3 } (Postgres) or just 3 (SQLite)
-            const row = rows[0];
-            const gameId = typeof row === 'object' ? row.id : row;
-
-            console.log("DEBUG: inserted rows =>", rows, "resolved gameId =>", gameId);
-
-            res.status(201).json({
-                message: 'New game started',
-                game_id: gameId,
-                start_time: Date.now()
-            });
-        })
-        .catch(err => {
-            console.error("ERROR inserting game state:", err);
-            res.status(500).json({ error: 'Failed to create new game' });
-        });
-});
-
-// Get the active game for this user
-app.get('/api/game/active', authenticateToken, (req, res) => {
-    db('game_states')
-        .where({ user_id: req.user.id, game_status: 'playing' })
-        .first()
-        .then(game => {
-            if (!game) return res.status(404).json({ error: 'No active game' });
-
-            // Parse JSON fields before sending
-            game.mine_positions = JSON.parse(game.mine_positions);
-            game.revealed_cells = JSON.parse(game.revealed_cells);
-            game.flagged_cells = JSON.parse(game.flagged_cells);
-
-            res.json(game);
-        })
-        .catch(err => res.status(500).json({ error: 'Failed to fetch active game' }));
-});
-
-// Get current game
-app.get('/api/game/current', authenticateToken, async (req, res) => {
-    try {
-        const game = await db('game_states')
-            .where({ user_id: req.user.id })
-            .orderBy('updated_at', 'desc')
-            .first();
-
-        if (!game) {
-            return res.status(404).json({ error: 'No active game found' });
-        }
-
-        res.json({
-            game_id: game.id,
-            level: game.level_id,
-            mine_count: game.mine_count,
-            mine_positions: JSON.parse(game.mine_positions),
-            revealed_cells: JSON.parse(game.revealed_cells),
-            flagged_cells: JSON.parse(game.flagged_cells),
-            hints: game.hints,
-            streak: game.streak,
-            game_status: game.game_status,
-        });
-    } catch (err) {
-        console.error(err);
-        res.status(500).json({ error: 'Failed to load game state' });
-    }
-});
-
-
-
-app.put('/api/game/update', authenticateToken, (req, res) => {
-    const { game_id, revealed_cells, flagged_cells, hints } = req.body;
-
-    if (!game_id || !Array.isArray(revealed_cells) || !Array.isArray(flagged_cells)) {
-        return res.status(400).json({ error: 'Game ID, revealed cells, and flagged cells are required' });
-    }
-
-    const updateData = {
-        revealed_cells: JSON.stringify(revealed_cells),
-        flagged_cells: JSON.stringify(flagged_cells),
-        updated_at: db.fn.now(),
-    };
-
-    if (typeof hints === 'number') {
-        updateData.hints = hints;
-    }
-
-    db('game_states')
-        .where({ id: game_id, user_id: req.user.id })
-        .update(updateData)
-        .then(changes => {
-            if (!changes || changes === 0) return res.status(404).json({ error: 'No active game found' });
-            res.json({ message: 'Game state updated successfully' });
-        })
-        .catch(err => {
-            console.error(err);
-            res.status(500).json({ error: 'Failed to update game state' });
-        });
-});
-
-
-// Get revealed cells
-app.get('/api/game/:gameId/revealed', authenticateToken, async (req, res) => {
-    const { gameId } = req.params;
-    const game = await db('game_states').where({ id: gameId, user_id: req.user.id }).first();
-    if (!game) return res.status(404).json({ error: 'Game not found' });
-    res.json({ revealed_cells: JSON.parse(game.revealed_cells) });
-});
-
-// Get flagged cells
-app.get('/api/game/:gameId/flagged', authenticateToken, async (req, res) => {
-    const { gameId } = req.params;
-    const game = await db('game_states').where({ id: gameId, user_id: req.user.id }).first();
-    if (!game) return res.status(404).json({ error: 'Game not found' });
-    res.json({ flagged_cells: JSON.parse(game.flagged_cells) });
-});
-
-
-// Finish game (transactional)
-app.post('/api/game/finish', authenticateToken, (req, res) => {
-    const { won, level, score, hints, streak } = req.body;
-    console.log("Finishing game with data:", req.body);
-
-    if (won === undefined || level === undefined || score === undefined) {
-        return res.status(400).json({ error: 'Won status, level, and score are required' });
-    }
-
-    db.transaction(trx => {
-        return trx('game_states')
-            .where({ user_id: req.user.id, game_status: 'playing' })
-            .first()
-            .then(game => {
-                if (!game) throw new Error('NO_ACTIVE_GAME');
-
-                // Update game_states
-                return trx('game_states')
-                    .where({ id: game.id })
-                    .update({
-                        game_status: won ? 'won' : 'lost',
-                        hints: hints || 0,
-                        streak: streak || 0,
-                        updated_at: db.fn.now()
-                    })
-                    .then(() => {
-                        if (!won) return { score_id: null, is_new_record: false };
-
-                        // Check if user has existing score
-                        return trx('scores')
-                            .where({ user_id: req.user.id })
-                            .first()
-                            .then(existingScore => {
-                                if (!existingScore) {
-                                    // No existing score, insert new
-                                    return trx('scores')
-                                        .insert({
-                                            user_id: req.user.id,
-                                            score,
-                                            level,
-                                            created_at: db.fn.now(),
-                                            updated_at: db.fn.now()
-                                        })
-                                        .returning('id')
-                                        .then(rows => ({
-                                            score_id: rows[0]?.id || rows[0], // Handle different return formats
-                                            is_new_record: true
-                                        }));
-                                }
-
-                                // Always update with the new cumulative score and level
-                                return trx('scores')
-                                    .where({ user_id: req.user.id })
-                                    .update({
-                                        score,
-                                        level,
-                                        updated_at: db.fn.now()
-                                    })
-                                    .then(() => ({
-                                        score_id: existingScore.id,
-                                        is_new_record: score > existingScore.score
-                                    }));
-                            });
-                    });
-            });
-    })
-        .then(result => {
-            console.log('✅ Game finished successfully:', result);
-            res.json({
-                message: 'Game finished successfully',
-                won,
-                score_id: result.score_id,
-                score,
-                level,
-                is_new_record: result.is_new_record
-            });
-        })
-        .catch(err => {
-            console.error('❌ Error finishing game:', err);
-            if (err.message === 'NO_ACTIVE_GAME') {
-                return res.status(404).json({ error: 'No active game found' });
-            }
-            res.status(500).json({ error: 'Server error', details: err.message });
-        });
-});
-
-// Get user's current score
-app.get('/api/user/score', authenticateToken, (req, res) => {
-    db('scores')
-        .where({ user_id: req.user.id })
-        .first()
-        .then(score => {
-            if (!score) {
-                return res.json({ score: 0, level: 0 });
-            }
-            res.json({
-                score: score.score || 0,
-                level: score.level || 0
-            });
-        })
-        .catch(err => {
-            console.error(err);
-            res.status(500).json({ error: 'Server error' });
-        });
-});
-
-
-
-// User's scores
-app.get('/api/scores/user', authenticateToken, (req, res) => {
-    const limit = parseInt(req.query.limit || '50');
-    const offset = parseInt(req.query.offset || '0');
-
-    db('scores')
-        .select('id', 'mine_count', 'level_id', 'created_at')
-        .where({ user_id: req.user.id })
-        .orderBy('created_at', 'desc')
-        .limit(limit)
-        .offset(offset)
-        .then(scores => res.json(scores))
-        .catch(err => res.status(500).json({ error: 'Server error' }));
-});
-
-
-app.get('/api/leaderboard', async (req, res) => {
-    try {
-        const { level } = req.query; // Changed from level_id to match your schema
-        const limit = parseInt(req.query.limit || '50');
-
-        // Build query to aggregate total scores per user
-        let query = db('scores as s')
-            .join('users as u', 's.user_id', 'u.id')
-            .select(
-                'u.username',
-                'u.country_flag',
-                db.raw('COALESCE(SUM(s.score), 0) as total_score'),
-                db.raw('COUNT(s.id) as games_played')
-            )
-            .groupBy('u.id', 'u.username', 'u.country_flag');
-
-        // Optional: filter by level if provided
-        if (level) {
-            query = query.where('s.level', level);
-        }
-
-        const results = await query
-            .orderBy('total_score', 'desc')
-            .limit(limit);
-
-        // Return country_flag code directly - Flutter will handle mapping
-        const leaderboard = results.map(row => ({
-            username: row.username,
-            total_score: parseInt(row.total_score) || 0,
-            country_flag: row.country_flag || 'international',
-            games_played: parseInt(row.games_played) || 0
-        }));
-
-        res.json(leaderboard);
-    } catch (err) {
-        console.error('Leaderboard error:', err);
-        res.status(500).json({ error: 'Failed to load leaderboard' });
-    }
-});
-
-// Get user statistics
-app.get('/api/user/stats', authenticateToken, async (req, res) => {
-    try {
-        const gamesPlayed = await db('game_states')
-            .where({ user_id: req.user.id })
-            .whereIn('game_status', ['won', 'lost'])
-            .count('* as count')
-            .first();
-
-        const gamesWon = await db('game_states')
-            .where({ user_id: req.user.id, game_status: 'won' })
-            .count('* as count')
-            .first();
-
-        const totalScore = await db('scores')
-            .where({ user_id: req.user.id })
-            .sum('score as total')
-            .first();
-
-        const played = parseInt(gamesPlayed?.count || '0');
-        const won = parseInt(gamesWon?.count || '0');
-        const total = parseInt(totalScore?.total || '0');
-
-        res.json({
-            games_played: played,
-            games_won: won,
-            total_score: total,
-        });
-    } catch (err) {
-        console.error(err);
-        res.status(500).json({ error: 'Failed to load stats' });
-    }
-});
-
-
-// Error handling middleware
-app.use((err, req, res, next) => {
-    console.error('Unhandled error:', err.stack || err);
-    res.status(500).json({ error: 'Something went wrong!' });
-});
-
-// 404
-app.use((req, res) => res.status(404).json({ error: 'Route not found' }));
-
-// Graceful shutdown
-process.on('SIGINT', () => {
-    console.log('\nShutting down gracefully...');
-    Promise.all([
-        db.destroy().catch(err => console.error('Error closing db', err)),
-        redis.quit().catch(err => console.error('Error closing redis', err))
-    ]).then(() => process.exit(0));
-});
-
-// Catch-all for undefined API routes
-app.use((req, res) => {
-    res.status(404).json({ error: "Route not found" });
-});
-
-// Global error handler (optional but recommended)
-app.use((err, req, res, next) => {
-    console.error("Unhandled error:", err);
-    res.status(500).json({ error: "Internal Server Error" });
-});
-
-app.listen(PORT, () => {
-    console.log(`🚀 Mine Master API running at http://localhost:${PORT}`);
-});
+// Public routes
+app.use(healthRoutes);
+app.use(flagsRoutes);
+app.use('/api/auth', authRoutes);
+
+// Protected routes (if you want, you can apply middleware per-route instead)
+app.use('/api', authenticateToken, userRoutes);
+app.use('/api', authenticateToken, gameRoutes);
+app.use('/api', leaderboardRoutes); // leaderboard might be public; keep as-is
+
+app.use(notFoundHandler);
+app.use(errorHandler);
+
+app.listen(PORT, () => console.log(`🚀 Mine Master API running at http://localhost:${PORT}`));
